+++ src/test/webapp/codecept.video.conf.js
@@ -1,4 +1,11 @@
 const baseConfig = require("./codecept.conf.js").config;
+const { getVideoSettings } = require("./helpers/video_settings.js");
+
+const videoSettings = getVideoSettings();
+
+console.log("videoSize=", `${videoSettings.width}x${videoSettings.height}`);
+console.log("videoViewport=", `${videoSettings.viewportWidth}x${videoSettings.viewportHeight}`);
+console.log("videoZoom=", videoSettings.zoom);
 
 exports.config = {
   ...baseConfig,
@@ -7,7 +14,19 @@ exports.config = {
     ...baseConfig.helpers,
     Playwright: {
       ...baseConfig.helpers.Playwright,
-      require: "./helpers/video_playwright_helper.js"
+      require: "./helpers/video_playwright_helper.js",
+      windowSize: `${videoSettings.width}x${videoSettings.height}`,
+      video: true,
+      keepVideoForPassedTests: true,
+      recordVideo: {
+        size: {
+          width: videoSettings.width,
+          height: videoSettings.height
+        }
+      }
+    },
+    VideoHelper: {
+      require: "./helpers/video_helper.js"
     },
     AudioHelper: {
       require: "./helpers/audio_helper.js",
